fix the header file

+++ Get_Next_Line/get_next_line.h
@@ -25,7 +25,7 @@ char    *get_next_line(int fd);
 int     ft_strlen(char *str);
 char    *ft_substr(char *src, int from, int to, int flag);
 char    *ft_strncpy(char *dst, char *src, int n);
-int     ft_strchr(char *src, int *ptr);
+int     ft_strchr(char *src, ssize_t *ptr);
 char    *ft_strjoin(char *origin, char *added);
 
 #endif
